pyproject.toml
[project]
name = "rgpv-live"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.11",
    "djangorestframework>=3.16.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-core>=1.2.14",
    "langchain-openai>=1.1.10",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
]
